requirements.txt
# --- Backend web ---
fastapi==0.110.0
uvicorn[standard]==0.29.0

# --- Templates e arquivos estáticos ---
jinja2==3.1.3
aiofiles==23.2.1

# --- Azure AI ---
azure-ai-formrecognizer==3.3.2
azure-core==1.30.1

# --- Manipulação de PDFs (caso use localmente) ---
PyMuPDF==1.24.2
pdfplumber==0.10.3

# --- Uploads, variáveis de ambiente e formulário ---
python-multipart==0.0.9
python-dotenv==1.0.1

# --- Logging / debug (opcional) ---
loguru==0.7.2

# --- Regex avançado (opcional) ---
regex==2023.12.25

# --- Compatibilidade com VS Code e Pylance ---
requests==2.31.0
